Refactored SetProps to handle unmounted components
The standalone version of SetProps must be able to
decide skipping updating a component, if it is
already unmounted otherwise React throws an error.

The current implementation basically handles not
mounted components, as if they don't exist. Props
send to the noflo component will thus be stored
and applied to the next component arriving.